# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/public/login": {
            "post": {
                "description": "Authenticates a user and returns an access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "LoginUser",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/commands.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or unknown provider",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "422": {
                        "description": "Unprocessable input (validation failed)",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/public/logout": {
            "post": {
                "description": "Logs out the authenticated user.\nExample success response: {\"success\": true}\nExample error response: {\"success\": false, \"error\": {\"code\": \"USER_NOT_FOUND\", \"message\": \"User not found\", \"status\": \"Not Found\"}}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logout completed successfully",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or unknown provider",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "422": {
                        "description": "Unprocessable input (validation failed)",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/public/register": {
            "post": {
                "description": "Handles user registration by parsing the request body and invoking the registration command.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "RegisterUser request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/commands.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or unknown provider",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "422": {
                        "description": "Unprocessable input (validation failed)",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "commands.LoginUser": {
            "type": "object",
            "required": [
                "password",
                "user_name"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "commands.RegisterUser": {
            "type": "object",
            "required": [
                "avatar_identifier",
                "email",
                "first_name",
                "last_name",
                "password",
                "user_name"
            ],
            "properties": {
                "avatar_identifier": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "minLength": 3
                },
                "first_name": {
                    "type": "string",
                    "minLength": 3
                },
                "last_name": {
                    "type": "string",
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "minLength": 3
                },
                "user_name": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "http.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code (from error ID)",
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status text",
                    "type": "string"
                }
            }
        },
        "http.ResponseResult": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/http.HTTPError"
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}